+++ src/package_index.rs
@@ -65,11 +65,16 @@ impl PackageIndex {
         };
 
         index.update()?;
+
         Ok(index)
     }
 
     pub fn url(&self) -> &Url {
         &self.url
+    }
+
+    pub fn path(&self) -> &PathBuf {
+        &self.path
     }
 
     pub fn new_temp(index_url: &Url, access_token: Option<String>) -> anyhow::Result<Self> {
@@ -177,8 +182,10 @@ impl PackageIndex {
                     .into_iter::<Manifest>()
                     .collect();
 
-            let versions = manifest_stream
+            let mut versions = manifest_stream
                 .with_context(|| format!("could not parse package index entry for {}", name))?;
+
+            versions.sort_by(|a, b| b.package.version.cmp(&a.package.version));
 
             let metadata = Arc::new(PackageMetadata { versions });
             package_cache.insert(name.clone(), Arc::clone(&metadata));
@@ -252,7 +259,7 @@ impl PackageIndex {
     }
 }
 
-#[derive(Default)]
+#[derive(Default, Serialize)]
 pub struct PackageMetadata {
     pub versions: Vec<Manifest>,
 }